modify responses

+++ controllers/controller.js
@@ -6,7 +6,7 @@ const Database = require("../models/Database")
 exports.getImage = (req, res) => {
     const db = new Database()
     db.getImage(req, (result) => {
-        if (result == "Error")
+        if ("error" in result)
             res.send("Error")
         else {
             result.file = `/${parameter.fileupload.storage}/` + result.file
@@ -20,15 +20,13 @@ exports.getImage = (req, res) => {
 exports.getImages = (req, res) => {
     const db = new Database()
     db.getImages((result) => {
-        if (result == "Error")
-            res.send("Error")
-        else {
+        if (!"error" in result) {
             result.forEach(e => {
                 e.file = `/${parameter.fileupload.storage}/` + e.file
             })
-            res.send(result)
-            db.end()
         }
+        res.send(result)
+        db.end()
     })
 }
 
@@ -39,6 +37,8 @@ exports.addImage = (req, res) => {
     else {
         const db = new Database()
         db.addImage(req, (result) => {
+            if ("error" in result)
+                fs.unlinkSync(path.resolve(`./${parameter.fileupload.storage}/${req.file.filename}`))
             res.send(result)
             db.end()
         })
@@ -51,8 +51,9 @@ exports.deleteImage = (req, res) => {
         if (result == "Error")
             res.send("Error")
         else {
-            fs.unlinkSync(path.resolve(`./${parameter.fileupload.storage}/${result.file}`))
             db.deleteImage(req, (result) => {
+                if (!"error" in result)
+                    fs.unlinkSync(path.resolve(`./${parameter.fileupload.storage}/${result.file}`))
                 res.send(result)
                 db.end()
             })
